Move only class from inlining package to interpreter
