+++ src/som/compiler/MethodBuilder.java
@@ -44,9 +44,9 @@ import som.compiler.ProgramDefinitionError.SemanticDefinitionError;
 import som.compiler.Variable.Argument;
 import som.compiler.Variable.Internal;
 import som.compiler.Variable.Local;
-import som.inlining.InliningVisitor;
 import som.interpreter.LexicalScope.MethodScope;
 import som.interpreter.LexicalScope.MixinScope;
+import som.interpreter.InliningVisitor;
 import som.interpreter.Method;
 import som.interpreter.SNodeFactory;
 import som.interpreter.nodes.ExpressionNode;
